requirements.txt
Flask
flask-cors
google-generativeai==0.7.2
